pyproject.toml
[project]
name = "care_digit_integration"
version = "0.1.0"
description = "Plugin for integration of DIGIT with CARE"
readme = "README.md"
authors = [
  {name = "Open Health Care Network", email = "support@ohc.network"}
]
maintainers = [
  {name = "Open Health Care Network", email = "support@ohc.network"}
]
classifiers = [
    # TODO
]
license = {text = "MIT"}
dependencies = [
    "django",
    "celery",
    "djangorestframework",
    "requests",
]
requires-python = ">= 3.10"

[project.optional-dependencies]
test = [
    "ruff",  # linting
]

[project.urls]
bugs = "https://github.com/ohcnetwork/care_digit_integration/issues"
changelog = "https://github.com/ohcnetwork/care_digit_integration/blob/master/changelog.md"
homepage = "https://github.com/ohcnetwork/care_digit_integration"


[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.uv]
package = true
